requirements.txt
dash
dash-bootstrap-components
pandas
rapidfuzz
Flask-Cors
requests
python-dotenv
numpy
plotly
Flask-Caching
waitress
setuptools